Extract useReelController from SpinReel
SpinReel was 264 lines and held 9 refs/MotionValues plus an 80-line
useImperativeHandle that combined a tween runner, a Promise resolver
chain and elapsed-time tracking. Move that imperative API and its
backing state into a hook. SpinReel keeps the MotionValue, the virtual
window, the visible-slot rendering and the phase/haptic effects, and
just forwards the controller through useImperativeHandle.

The centerOffset stale-closure that the old code patched with a manual
ref+effect now lives inside the hook as a single ref, fed from the
prop. The skip-during-active-tween invariant on resize is preserved.
